Move PluginProvider implementation into _plugin
